Refactor to use article & page generator write signals
Modify the plugin to work with the `article_generator_write_article` and
`page_generator_write_page` signals instead of 'content_object_init'.

Now the plugin injects code as each article and page are being written,
rather than during content initialization. This change allows for
greater flexibility and control over where the code gets injected.
In addition, user-configurable settings have been added to
enable/disable injection in pages and/or articles.

Note: The `content_object_init` signal was not suitable for injecting
into `<head>` or other parts of the HTML document that are not included
in the content object. The newly chosen signals provide access to the
full generated HTML of each page and article, overcoming this limitation.